first stable version on Prod

+++ src/main/scala/com/nbnco/csa/analysis/copper/sdc/data/SdcDataHistorical.scala
@@ -13,15 +13,15 @@ package com.nbnco.csa.analysis.copper.sdc.data
   * @param reTransUs    xdslFarEndChannelPreviousIntervalRetransmDtuCounterUS
   * @param reTransDs    xdslChannelPreviousIntervalRetransmDtuCounterDS
   */
-case class SdcDataHistorical(var ses: Int,
-                             var uas: Int,
-                             var lprFe: Int,
-                             var sesFe: Int,
-                             var unCorrDtuDs: Int,
-                             var unCorrDtuUs: Int,
-                             var reInit: Int,
-                             var reTransUs : Int,
-                             var reTransDs: Int
+case class SdcDataHistorical(var ses: Long,
+                             var uas: Long,
+                             var lprFe: Long,
+                             var sesFe: Long,
+                             var unCorrDtuDs: Long,
+                             var unCorrDtuUs: Long,
+                             var reInit: Long,
+                             var reTransUs : Long,
+                             var reTransDs: Long
                             ) extends SdcDataBase {
 	def toMap: Map[String, Any] = {
 		Map (
